Using visitors for npz
