+++ src/nn/impl_module_for_tuples.rs
@@ -5,7 +5,7 @@ use super::module::{BuildModule, BuildOnDevice, Module, ModuleMut, OnDevice, ToD
 macro_rules! tuple_impls {
     ([$($name:ident),+] [$($idx:tt),+], $last:ident, [$($rev_tail:ident),+]) => {
         impl<E: Dtype, D: DeviceStorage, $($name: TensorCollection<E, D>),+> TensorCollection<E, D> for ($($name,)+) {
-            fn iter_tensors<V: ModuleWalker<Self, E, D>>(visitor: &mut V) -> Result<(), D::Err> {
+            fn iter_tensors<V: ModuleWalker<Self, E, D>>(visitor: &mut V) -> Result<(), V::Err> {
                 $(visitor.visit_module(|s| &s.$idx, |s| &mut s.$idx, &std::format!("{}", $idx))?;)+
                 Ok(())
             }
