+++ src/nn/transformer/mod.rs
@@ -102,7 +102,7 @@ where
     E: Dtype + Float + SampleUniform,
     D: Device<E>,
 {
-    fn iter_tensors<V: ModuleWalker<Self, E, D>>(visitor: &mut V) -> Result<(), <D>::Err> {
+    fn iter_tensors<V: ModuleWalker<Self, E, D>>(visitor: &mut V) -> Result<(), V::Err> {
         visitor.visit_module(|s| &s.encoder, |s| &mut s.encoder, "encoder")?;
         visitor.visit_module(|s| &s.decoder, |s| &mut s.decoder, "decoder")
     }
