+++ src/optim/optimizer.rs
@@ -95,7 +95,7 @@ pub trait Optimizer<M, D: DeviceStorage, E: Dtype> {
 /// Represents something that can be updated with a [ParamUpdater].
 pub trait GradientUpdate<E: Dtype, D: DeviceStorage>: TensorCollection<E, D> {
     /// Updates self given the [ParamUpdater].
-    fn update<V: VisitTensorMut<E, D>>(&mut self, updater: &mut V) -> Result<(), D::Err> {
+    fn update<V: VisitTensorMut<E, D>>(&mut self, updater: &mut V) -> Result<(), V::Err> {
         Self::iter_tensors(&mut RecursiveWalker {
             m: self,
             f: updater,
